requirements.txt
boto3==1.14.28
botocore==1.17.28
certifi==2020.6.20
chardet==3.0.4
docutils==0.15.2
fire==0.3.1
idna==2.10
jmespath==0.10.0
numpy==1.19.1
opencv-python==4.3.0.36
pdf2image==1.13.1
Pillow==7.2.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.18
termcolor==1.1.0
urllib3==1.25.10
